2023/2/12 -- Read description
I edited the Third Person Controller as well as the Player State Manager so it could actually do the shooting animation, and I have to wait for one of the errors for the camera to be fixed as well (NullException occurs when the camera is trying to mark the Far Camera to SetActive = true). Temporarily I had to comment out that line of code in the camera's script to see if the ball would shoot at all, and that seems to be the only roadblock. Later this week, I will have to fix the logic behind the animator as well as probably replacing the entire thing with maybe a State Machine.