Read description
Allowed 8-way traversal of the character, stopped the buggy dribble animation, fixed the inverted controls when the game has played for an extended amount of time, and fixed a bug where, on rare occasions, the character would lock up completely and does not do any animations. I found that it would freeze at the end of the Shooting animation because it was the only animation that was overlooked having an [Exit time], where, once the animation reaches its [Exit], the Controller does not know what to do especially when multiple parameters are changing at the same time and the controller cannot catch up to the rapid changes of those parameters. The below video is demonstrating absurd movements on the character to make sure the animations dont freeze as Parameters change.