Main Menu/Credits current progress
I was able to create a new scene for the credits and working on improving it so that it was working properly compared to my last iteration. I added the settings menu with small but effective settings such as a fullscreen button, resolution dropdown, and volume slider. Much more work to be done but some good progress!